Extract out fillOutForm helper
