+++ test/test-va.js
@@ -2,6 +2,7 @@ const puppeteer = require('puppeteer');
 const chai = require('chai');
 chai.use(require('chai-string'));
 const assert = chai.assert;
+const fillOutForm = require('./helpers').fillOutForm;
 
 describe('VA SNAP prescreener', () => {
     before(async () => {
@@ -19,28 +20,6 @@ describe('VA SNAP prescreener', () => {
         await page.close();
         await browser.close();
     });
-
-    const fillOutForm = async (steps) => {
-        await page.select('#household_size', steps['household_size']);
-
-        if (steps['household_includes_elderly_or_disabled']) {
-            await page.click('label[for="input__household_includes_elderly_or_disabled_true"]');
-        } else {
-            await page.click('label[for="input__household_includes_elderly_or_disabled_false"]');
-        }
-
-        if (steps['all_citizens']) {
-            await page.click('label[for="input__all_citizens_question_true"]');
-        } else {
-            await page.click('label[for="input__all_citizens_question_false"]');
-
-        }
-        await page.type('#monthly_job_income', steps['monthly_job_income']);
-        await page.type('#monthly_non_job_income', steps['monthly_non_job_income']);
-        await page.type('#resources', steps['monthly_job_income']);
-
-        await page.click('#prescreener-form-submit');
-    }
 
     it('shows the correct results HTML for a 1-person eligible household', async () => {
         fillOutForm({
